# Improve packing performance on large binary data.