This improves performance for packing large binary data.

On the benchmark at `example/binary_benchmark.dart`, I achieved about a 2.5x to 2.7x speed increase in packing performance when packing 60MB.

This reduces the time needed to fully pack 60MB from 530ms to 197ms.

Before these changes:
```
PS C:\Code\MsgPack> dart .\example\binary_benchmark.dart
Took 551.489ms to encode.
Took 0.194ms to decode.
Took 530.689ms to encode.
Took 0.194ms to decode.
Took 500.7ms to encode.
Took 0.239ms to decode.
Took 526.304ms to encode.
Took 0.276ms to decode.
Took 545.068ms to encode.
Took 0.211ms to decode.
Took an average of 530.85ms to encode.
Took an average of 0.22ms to decode.
```

After these changes:
```
PS C:\Code\MsgPack> dart .\example\binary_benchmark.dart
Took 188.548ms to encode.
Took 0.17ms to decode.
Took 201.682ms to encode.
Took 0.229ms to decode.
Took 210.269ms to encode.
Took 0.208ms to decode.
Took 199.06ms to encode.
Took 0.177ms to decode.
Took 187.601ms to encode.
Took 0.182ms to decode.
Took an average of 197.43ms to encode.
Took an average of 0.19ms to decode.
```

This also adds a way to reuse buffers without recreation, however it is not enabled unless you call a special method.

The algorithm used to pack binaries now tries really hard to use the Dart intrinsic for memcpy() like performance on `Uint8List`.